+++ src/main/java/odin/parser/command/ManipulateCommand.java
@@ -23,9 +23,9 @@ public abstract class ManipulateCommand implements Command {
             throw new WrongFormatException(String.format("'%s' command must be followed by an integer.", this.getCommandName()));
         }
         this.taskList = taskList;
-        this.idx = (Integer) Integer.parseInt(tokens.get(0));
+        this.idx = Integer.parseInt(tokens.get(0));
         if (this.idx <= 0 || this.idx > taskList.getSize()) {
-            throw new WrongFormatException(String.format("The task index you speak of is incorrect. There are tasks numbered 1 through %d.", (Integer) taskList.getSize()));
+            throw new WrongFormatException(String.format("The task index you speak of is incorrect. There are tasks numbered 1 through %d.", taskList.getSize()));
         }
         this.manipulate();
     }
